Add service contracts
