+++ packages/contracts/source/contracts/cli/index.ts
@@ -1,1 +1,4 @@
+export * from "./application.js"
 export * from "./cli.js"
+export * from "./paths.js"
+export * from "./services.js"
