+++ packages/contracts/source/contracts/cli/paths.ts
@@ -1,0 +1,1 @@
+export type { Paths } from "env-paths";
